feat(api): provide a flavor of api using stdbool
API type: RimeApi_stdbool
API entry function: rime_get_api_stdbool
