+++ src/rime_api.h
@@ -30,22 +30,27 @@ extern "C" {
 #define RIME_API
 #endif /* _WIN32 */
 
-typedef uintptr_t RimeSessionId;
+#ifndef RIME_DEPRECATED
+#define RIME_DEPRECATED RIME_API
+#endif
 
-typedef int Bool;
+#ifndef RIME_FLAVORED
+#define RIME_FLAVORED(name) name
+#endif
+
+#ifndef Bool
+#define Bool int
+#endif
 
 #ifndef False
 #define False 0
 #endif
+
 #ifndef True
 #define True 1
 #endif
 
-//! Define the max number of candidates
-/*!
- *  \deprecated There is no limit to the number of candidates in RimeMenu
- */
-#define RIME_MAX_NUM_CANDIDATES 10
+typedef uintptr_t RimeSessionId;
 
 // Version control
 #define RIME_STRUCT_INIT(Type, var) \
@@ -60,6 +65,10 @@ typedef int Bool;
 #define RIME_STRUCT(Type, var) \
   Type var = {0};              \
   RIME_STRUCT_INIT(Type, var);
+
+// tests that member is a non-null pointer in self-versioned struct *p.
+#define RIME_PROVIDED(p, member) \
+  ((p) && RIME_STRUCT_HAS_MEMBER(*(p), (p)->member) && (p)->member)
 
 //! For passing pointer to capnproto builder as opaque pointer through C API.
 #define RIME_PROTO_BUILDER void
@@ -125,7 +134,7 @@ typedef struct {
   int num_candidates;
   RimeCandidate* candidates;
   char* select_keys;
-} RimeMenu;
+} RIME_FLAVORED(RimeMenu);
 
 /*!
  *  Should be initialized by calling RIME_STRUCT_INIT(Type, var);
@@ -139,20 +148,20 @@ typedef struct rime_commit_t {
 /*!
  *  Should be initialized by calling RIME_STRUCT_INIT(Type, var);
  */
-typedef struct rime_context_t {
+typedef struct RIME_FLAVORED(rime_context_t) {
   int data_size;
   // v0.9
   RimeComposition composition;
-  RimeMenu menu;
+  RIME_FLAVORED(RimeMenu) menu;
   // v0.9.2
   char* commit_text_preview;
   char** select_labels;
-} RimeContext;
+} RIME_FLAVORED(RimeContext);
 
 /*!
  *  Should be initialized by calling RIME_STRUCT_INIT(Type, var);
  */
-typedef struct rime_status_t {
+typedef struct RIME_FLAVORED(rime_status_t) {
   int data_size;
   // v0.9
   char* schema_id;
@@ -164,7 +173,7 @@ typedef struct rime_status_t {
   Bool is_simplified;
   Bool is_traditional;
   Bool is_ascii_punct;
-} RimeStatus;
+} RIME_FLAVORED(RimeStatus);
 
 typedef struct rime_candidate_list_iterator_t {
   void* ptr;
@@ -195,32 +204,11 @@ typedef struct rime_schema_list_t {
   RimeSchemaListItem* list;
 } RimeSchemaList;
 
-typedef void (*RimeNotificationHandler)(void* context_object,
-                                        RimeSessionId session_id,
-                                        const char* message_type,
-                                        const char* message_value);
-
 typedef struct rime_string_slice_t {
   const char* str;
   size_t length;
 } RimeStringSlice;
 
-// Setup
-
-/*!
- *  Call this function before accessing any other API.
- */
-RIME_API void RimeSetup(RimeTraits* traits);
-
-/*!
- *  Pass a C-string constant in the format "rime.x"
- *  where 'x' is the name of your application.
- *  Add prefix "rime." to ensure old log files are automatically cleaned.
- *  \deprecated Use RimeSetup() instead.
- */
-RIME_API void RimeSetupLogging(const char* app_name);
-
-//! Receive notifications
 /*!
  * - on loading schema:
  *   + message_type="schema", message_value="luna_pinyin/Luna Pinyin"
@@ -236,159 +224,10 @@ RIME_API void RimeSetupLogging(const char* app_name);
  *   every time an event occurs in librime, until RimeFinalize() is called.
  *   when handler is NULL, notification is disabled.
  */
-RIME_API void RimeSetNotificationHandler(RimeNotificationHandler handler,
-                                         void* context_object);
-
-// Entry and exit
-
-RIME_API void RimeInitialize(RimeTraits* traits);
-RIME_API void RimeFinalize(void);
-
-RIME_API Bool RimeStartMaintenance(Bool full_check);
-
-//! \deprecated Use RimeStartMaintenance(full_check = False) instead.
-RIME_API Bool RimeStartMaintenanceOnWorkspaceChange(void);
-RIME_API Bool RimeIsMaintenancing(void);
-RIME_API void RimeJoinMaintenanceThread(void);
-
-// Deployment
-
-RIME_API void RimeDeployerInitialize(RimeTraits* traits);
-RIME_API Bool RimePrebuildAllSchemas(void);
-RIME_API Bool RimeDeployWorkspace(void);
-RIME_API Bool RimeDeploySchema(const char* schema_file);
-RIME_API Bool RimeDeployConfigFile(const char* file_name,
-                                   const char* version_key);
-
-RIME_API Bool RimeSyncUserData(void);
-
-// Session management
-
-RIME_API RimeSessionId RimeCreateSession(void);
-RIME_API Bool RimeFindSession(RimeSessionId session_id);
-RIME_API Bool RimeDestroySession(RimeSessionId session_id);
-RIME_API void RimeCleanupStaleSessions(void);
-RIME_API void RimeCleanupAllSessions(void);
-
-// Input
-
-RIME_API Bool RimeProcessKey(RimeSessionId session_id, int keycode, int mask);
-/*!
- * return True if there is unread commit text
- */
-RIME_API Bool RimeCommitComposition(RimeSessionId session_id);
-RIME_API void RimeClearComposition(RimeSessionId session_id);
-
-// Output
-
-RIME_API Bool RimeGetCommit(RimeSessionId session_id, RimeCommit* commit);
-RIME_API Bool RimeFreeCommit(RimeCommit* commit);
-RIME_API Bool RimeGetContext(RimeSessionId session_id, RimeContext* context);
-RIME_API Bool RimeFreeContext(RimeContext* context);
-RIME_API Bool RimeGetStatus(RimeSessionId session_id, RimeStatus* status);
-RIME_API Bool RimeFreeStatus(RimeStatus* status);
-
-// Accessing candidate list
-RIME_API Bool RimeCandidateListBegin(RimeSessionId session_id,
-                                     RimeCandidateListIterator* iterator);
-RIME_API Bool RimeCandidateListNext(RimeCandidateListIterator* iterator);
-RIME_API void RimeCandidateListEnd(RimeCandidateListIterator* iterator);
-RIME_API Bool RimeCandidateListFromIndex(RimeSessionId session_id,
-                                         RimeCandidateListIterator* iterator,
-                                         int index);
-RIME_API Bool RimeSelectCandidate(RimeSessionId session_id, size_t index);
-RIME_API Bool RimeSelectCandidateOnCurrentPage(RimeSessionId session_id,
-                                               size_t index);
-RIME_API Bool RimeDeleteCandidate(RimeSessionId session_id, size_t index);
-RIME_API Bool RimeDeleteCandidateOnCurrentPage(RimeSessionId session_id,
-                                               size_t index);
-
-// Runtime options
-
-RIME_API void RimeSetOption(RimeSessionId session_id,
-                            const char* option,
-                            Bool value);
-RIME_API Bool RimeGetOption(RimeSessionId session_id, const char* option);
-
-RIME_API void RimeSetProperty(RimeSessionId session_id,
-                              const char* prop,
-                              const char* value);
-RIME_API Bool RimeGetProperty(RimeSessionId session_id,
-                              const char* prop,
-                              char* value,
-                              size_t buffer_size);
-
-RIME_API Bool RimeGetSchemaList(RimeSchemaList* schema_list);
-RIME_API void RimeFreeSchemaList(RimeSchemaList* schema_list);
-RIME_API Bool RimeGetCurrentSchema(RimeSessionId session_id,
-                                   char* schema_id,
-                                   size_t buffer_size);
-RIME_API Bool RimeSelectSchema(RimeSessionId session_id, const char* schema_id);
-
-// Configuration
-
-// <schema_id>.schema.yaml
-RIME_API Bool RimeSchemaOpen(const char* schema_id, RimeConfig* config);
-// <config_id>.yaml
-RIME_API Bool RimeConfigOpen(const char* config_id, RimeConfig* config);
-// access config files in user data directory, eg. user.yaml and
-// installation.yaml
-RIME_API Bool RimeUserConfigOpen(const char* config_id, RimeConfig* config);
-RIME_API Bool RimeConfigClose(RimeConfig* config);
-RIME_API Bool RimeConfigInit(RimeConfig* config);
-RIME_API Bool RimeConfigLoadString(RimeConfig* config, const char* yaml);
-// Access config values
-RIME_API Bool RimeConfigGetBool(RimeConfig* config,
-                                const char* key,
-                                Bool* value);
-RIME_API Bool RimeConfigGetInt(RimeConfig* config, const char* key, int* value);
-RIME_API Bool RimeConfigGetDouble(RimeConfig* config,
-                                  const char* key,
-                                  double* value);
-RIME_API Bool RimeConfigGetString(RimeConfig* config,
-                                  const char* key,
-                                  char* value,
-                                  size_t buffer_size);
-RIME_API const char* RimeConfigGetCString(RimeConfig* config, const char* key);
-RIME_API Bool RimeConfigSetBool(RimeConfig* config,
-                                const char* key,
-                                Bool value);
-RIME_API Bool RimeConfigSetInt(RimeConfig* config, const char* key, int value);
-RIME_API Bool RimeConfigSetDouble(RimeConfig* config,
-                                  const char* key,
-                                  double value);
-RIME_API Bool RimeConfigSetString(RimeConfig* config,
-                                  const char* key,
-                                  const char* value);
-// Manipulate complex structures
-RIME_API Bool RimeConfigGetItem(RimeConfig* config,
-                                const char* key,
-                                RimeConfig* value);
-RIME_API Bool RimeConfigSetItem(RimeConfig* config,
-                                const char* key,
-                                RimeConfig* value);
-RIME_API Bool RimeConfigClear(RimeConfig* config, const char* key);
-RIME_API Bool RimeConfigCreateList(RimeConfig* config, const char* key);
-RIME_API Bool RimeConfigCreateMap(RimeConfig* config, const char* key);
-RIME_API size_t RimeConfigListSize(RimeConfig* config, const char* key);
-RIME_API Bool RimeConfigBeginList(RimeConfigIterator* iterator,
-                                  RimeConfig* config,
-                                  const char* key);
-RIME_API Bool RimeConfigBeginMap(RimeConfigIterator* iterator,
-                                 RimeConfig* config,
-                                 const char* key);
-RIME_API Bool RimeConfigNext(RimeConfigIterator* iterator);
-RIME_API void RimeConfigEnd(RimeConfigIterator* iterator);
-// Utilities
-RIME_API Bool RimeConfigUpdateSignature(RimeConfig* config, const char* signer);
-
-// Testing
-
-RIME_API Bool RimeSimulateKeySequence(RimeSessionId session_id,
-                                      const char* key_sequence);
-
-RIME_API Bool RimeSetInput(RimeSessionId session_id, const char* input);
-// Module
+typedef void (*RimeNotificationHandler)(void* context_object,
+                                        RimeSessionId session_id,
+                                        const char* message_type,
+                                        const char* message_value);
 
 /*!
  *  Extend the structure to publish custom data/functions in your specific
@@ -410,22 +249,10 @@ typedef struct rime_module_t {
 RIME_API Bool RimeRegisterModule(RimeModule* module);
 RIME_API RimeModule* RimeFindModule(const char* module_name);
 
-//! Run a registered task
-RIME_API Bool RimeRunTask(const char* task_name);
-
-//! \deprecated use RimeApi::get_shared_data_dir_s instead.
-RIME_API const char* RimeGetSharedDataDir(void);
-//! \deprecated use RimeApi::get_user_data_dir_s instead.
-RIME_API const char* RimeGetUserDataDir(void);
-//! \deprecated use RimeApi::get_sync_dir_s instead.
-RIME_API const char* RimeGetSyncDir(void);
-
-RIME_API const char* RimeGetUserId(void);
-
 /*! The API structure
  *  RimeApi is for rime v1.0+
  */
-typedef struct rime_api_t {
+typedef struct RIME_FLAVORED(rime_api_t) {
   int data_size;
 
   /*! setup
@@ -490,10 +317,12 @@ typedef struct rime_api_t {
 
   Bool (*get_commit)(RimeSessionId session_id, RimeCommit* commit);
   Bool (*free_commit)(RimeCommit* commit);
-  Bool (*get_context)(RimeSessionId session_id, RimeContext* context);
-  Bool (*free_context)(RimeContext* ctx);
-  Bool (*get_status)(RimeSessionId session_id, RimeStatus* status);
-  Bool (*free_status)(RimeStatus* status);
+  Bool (*get_context)(RimeSessionId session_id,
+                      RIME_FLAVORED(RimeContext) * context);
+  Bool (*free_context)(RIME_FLAVORED(RimeContext) * ctx);
+  Bool (*get_status)(RimeSessionId session_id,
+                     RIME_FLAVORED(RimeStatus) * status);
+  Bool (*free_status)(RIME_FLAVORED(RimeStatus) * status);
 
   // runtime options
 
@@ -672,13 +501,13 @@ typedef struct rime_api_t {
                                               size_t index);
 
   Bool (*change_page)(RimeSessionId session_id, Bool backward);
-} RimeApi;
+} RIME_FLAVORED(RimeApi);
 
 //! API entry
 /*!
  *  Acquire the version controlled RimeApi structure.
  */
-RIME_API RimeApi* rime_get_api(void);
+RIME_API RIME_FLAVORED(RimeApi) * RIME_FLAVORED(rime_get_api)(void);
 
 //! Clients should test if an api function is available in the current version
 //! before calling it.
