+++ src/rime_api_stdbool.cc
@@ -1,0 +1,4 @@
+#include "rime_api_stdbool.h"
+#include "rime_api.h"
+// implements stdbool flavored API
+#include "rime_api_impl.h"
